Split FunctionProvider to provider and functions
